LogLineDetails: extract into submodules
